WIP list index

+++ cpp/lexer.h
@@ -75,7 +75,7 @@ extern char ch;
 extern lextokens currenttoken;
 
 //Variable to store the last integer value encountered by the lexer
-extern int intvalue;
+extern double intvalue;
 
 //Used by main
 extern string lineInput;
